+++ common4j/src/main/com/microsoft/identity/common/java/cache/AbstractAccountCredentialCache.java
@@ -278,6 +278,26 @@ public abstract class AbstractAccountCredentialCache implements IAccountCredenti
     }
 
     /**
+     * The application identifier field for access token records previously included a SHA-1 signing certificate hash.
+     * This method identifies such tokens, as the application identifier should now contain a SHA-512 signing certificate hash.
+     * @param credential token record.
+     * @return true if an access token containing a SHA-1 app identifier hash; false otherwise.
+     */
+    public static boolean isSha1ApplicationIdentifierAccessToken(@NonNull final Credential credential) {
+        final String methodTag = TAG + ":isSha1ApplicationIdentifierAccessToken";
+        if (credential instanceof AccessTokenRecord) {
+            final AccessTokenRecord accessToken = (AccessTokenRecord) credential;
+            final String tokenAppIdentifier = accessToken.getApplicationIdentifier();
+            if (tokenAppIdentifier != null
+                    && applicationIdentifierContainsSha1(tokenAppIdentifier)) {
+                Logger.info(methodTag, "Identified old access token with app identifier containing SHA-1. This token shall be removed, and a new access token should be re-acquired with a SHA-512 app identifier.");
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
      * Can tell if an application identifier string (<package name>/<signing certificate hash>) contains a SHA-1 hash.
      * @param applicationIdentifier application identifier field string
      * @return true if contains SHA-1 hash; false otherwise.
